test(daemon): isolate remaining service workers
