+++ pyproject.toml
@@ -145,6 +145,7 @@ markers = [
   "daemon_bundle_refresh: enables the daemon supply-chain bundle refresh worker for focused tests",
   "daemon_headless_queue: enables request-triggered daemon cloud sync for focused tests",
   "daemon_headless_refresh: enables the daemon headless cloud sync worker for focused tests",
+  "daemon_service_workers: enables command queue and live request sync workers for focused tests",
   "slow: marks tests as slow (deselected by default; run with -m slow to include)",
 ]
